requirements.txt
langchain==0.2.12
langchain-core==0.2.29
langchain-openai==0.1.20
openai==1.40.1
python-dotenv
streamlit==1.37.1
langchain-community==0.2.11
langchain-groq==0.1.9
arxiv
wikipedia
duckduckgo_search
langchain-chroma==0.1.2
chromadb==0.5.5
pypdf
sentence-transformers
